refactor: admin 구조 수정
