+++ src/main/java/grepp/NBE5_6_2_Team03/api/controller/admin/dto/place/CountryResponse.java
@@ -1,4 +1,4 @@
-package grepp.NBE5_6_2_Team03.api.controller.admin.place.dto;
+package grepp.NBE5_6_2_Team03.api.controller.admin.dto.place;
 
 import grepp.NBE5_6_2_Team03.domain.admin.place.entity.Country;
 import java.util.List;
