+++ src/main/java/grepp/NBE5_6_2_Team03/domain/admin/place/PlaceService.java
@@ -1,7 +1,7 @@
 package grepp.NBE5_6_2_Team03.domain.admin.place;
 
-import grepp.NBE5_6_2_Team03.api.controller.admin.place.dto.PlaceResponse;
-import grepp.NBE5_6_2_Team03.api.controller.admin.place.dto.PlaceRequest;
+import grepp.NBE5_6_2_Team03.api.controller.admin.dto.place.PlaceResponse;
+import grepp.NBE5_6_2_Team03.api.controller.admin.dto.place.PlaceRequest;
 import grepp.NBE5_6_2_Team03.domain.admin.place.entity.Place;
 import grepp.NBE5_6_2_Team03.domain.admin.place.repository.PlaceRepository;
 import java.util.List;
